Masterproject-Sourcecode/backend_evaluation/requirements.txt
fastapi==0.115.12
json5==0.12.0
kubernetes==32.0.1
lancedb==0.21.2
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.58
langchain-huggingface==0.1.2
langchain-ollama==0.3.2
google-genai
ollama==0.4.7
pm4py==2.7.11.13
python-dotenv==1.0.0
#typing==3.11.0
uvicorn==0.23.2      matplotlib>=3.8.0
pandas>=2.0.0
numpy>=1.26.0